+++ trunk/src/scontrol/update_part.c
@@ -69,20 +69,20 @@ scontrol_update_part (int argc, char *argv[])
 		else if (strncasecmp(argv[i], "MaxNodes=", 9) == 0) {
 			if ((strcasecmp(&argv[i][9],"UNLIMITED") == 0) ||
 			    (strcasecmp(&argv[i][8],"INFINITE") == 0))
-				part_msg.max_nodes = (uint16_t) INFINITE;
+				part_msg.max_nodes = (uint32_t) INFINITE;
 			else
 				part_msg.max_nodes = 
-					(uint16_t) strtol(&argv[i][9], 
+					(uint32_t) strtol(&argv[i][9], 
 						(char **) NULL, 10);
 			update_cnt++;
 		}
 		else if (strncasecmp(argv[i], "MinNodes=", 9) == 0) {
 			if ((strcasecmp(&argv[i][9],"UNLIMITED") == 0) ||
 			    (strcasecmp(&argv[i][8],"INFINITE") == 0))
-				part_msg.min_nodes = (uint16_t) INFINITE;
+				part_msg.min_nodes = (uint32_t) INFINITE;
 			else
 				part_msg.min_nodes = 
-					(uint16_t) strtol(&argv[i][9], 
+					(uint32_t) strtol(&argv[i][9], 
 						(char **) NULL, 10);
 			update_cnt++;
 		}
